# [Graphbolt][CUDA] Optimize UVA by using CUB sort