The current implementation utilizes torch::sort which is slower than cub sort. This PR improves on the current code by swapping torch::sort with cub sort. The following throughput numbers are taken from #6590, the NVIDIA T4 results.

### Testing `USE_TORCH_SORT` on random datasets.

* params:  n_rows=100000, feat_size=100, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

|            |USE_TORCH_SORT=1|USE_TORCH_SORT=0|
|------------|------------|------------|
|USE_PERM=1  |       3.364|       4.454|
* params:  n_rows=100000, feat_size=100, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

|            |USE_TORCH_SORT=1|USE_TORCH_SORT=0|
|------------|------------|------------|
|USE_PERM=1  |    1605.974|    2421.417|
* params:  n_rows=100000, feat_size=100, num_indices=10000, feature_device=Device.Pinned, indices_device=Device.GPU

|            |USE_TORCH_SORT=1|USE_TORCH_SORT=0|
|------------|------------|------------|
|USE_PERM=1  |    3983.310|    4744.072|
* params:  n_rows=100000, feat_size=300, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

|            |USE_TORCH_SORT=1|USE_TORCH_SORT=0|
|------------|------------|------------|
|USE_PERM=1  |      11.516|      15.274|
* params:  n_rows=100000, feat_size=300, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

|            |USE_TORCH_SORT=1|USE_TORCH_SORT=0|
|------------|------------|------------|
|USE_PERM=1  |    3498.823|    4165.763|
* params:  n_rows=100000, feat_size=300, num_indices=10000, feature_device=Device.Pinned, indices_device=Device.GPU

|            |USE_TORCH_SORT=1|USE_TORCH_SORT=0|
|------------|------------|------------|
|USE_PERM=1  |    5421.254|    5651.394|

Additional results on A100 with #6644, note that when num_indices is 1 or 1000, the runtime is < 100us, meaning not enough work so throughput is not meaningful and we are measuring latency, even `num_indices=100000` sometimes is not enough work for the GPU to get full throughput.

Before this PR with #6644:
```
* params:  dataset=ogbn-products n_rows=2449029 feat_size=100 dtype=torch.float32 num_indices=518094 feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 9024, Throughput in MB/s: 21901


* params:  n_rows=2000000, feat_size=1, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 78, Throughput in MB/s: 0


* params:  n_rows=2000000, feat_size=1, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 69, Throughput in MB/s: 54


* params:  n_rows=2000000, feat_size=1, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 315, Throughput in MB/s: 1208


* params:  n_rows=2000000, feat_size=1, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 660, Throughput in MB/s: 5771


* params:  n_rows=2000000, feat_size=1, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 75, Throughput in MB/s: 0


* params:  n_rows=2000000, feat_size=1, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 71, Throughput in MB/s: 13


* params:  n_rows=2000000, feat_size=1, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 217, Throughput in MB/s: 437


* params:  n_rows=2000000, feat_size=1, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 434, Throughput in MB/s: 2195


* params:  n_rows=2000000, feat_size=4, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 79, Throughput in MB/s: 0


* params:  n_rows=2000000, feat_size=4, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 73, Throughput in MB/s: 206


* params:  n_rows=2000000, feat_size=4, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 487, Throughput in MB/s: 3130


* params:  n_rows=2000000, feat_size=4, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 1399, Throughput in MB/s: 10901


* params:  n_rows=2000000, feat_size=4, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 74, Throughput in MB/s: 0


* params:  n_rows=2000000, feat_size=4, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 71, Throughput in MB/s: 53


* params:  n_rows=2000000, feat_size=4, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 313, Throughput in MB/s: 1218


* params:  n_rows=2000000, feat_size=4, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 658, Throughput in MB/s: 5791


* params:  n_rows=2000000, feat_size=47, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 73, Throughput in MB/s: 2


* params:  n_rows=2000000, feat_size=47, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 68, Throughput in MB/s: 2617


* params:  n_rows=2000000, feat_size=47, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 1130, Throughput in MB/s: 15860


* params:  n_rows=2000000, feat_size=47, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 8065, Throughput in MB/s: 22230


* params:  n_rows=2000000, feat_size=47, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 74, Throughput in MB/s: 0


* params:  n_rows=2000000, feat_size=47, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 68, Throughput in MB/s: 653


* params:  n_rows=2000000, feat_size=47, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 538, Throughput in MB/s: 8323


* params:  n_rows=2000000, feat_size=47, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 3167, Throughput in MB/s: 14149


* params:  n_rows=2000000, feat_size=256, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 74, Throughput in MB/s: 13


* params:  n_rows=2000000, feat_size=256, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 96, Throughput in MB/s: 10119


* params:  n_rows=2000000, feat_size=256, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 4005, Throughput in MB/s: 24382


* params:  n_rows=2000000, feat_size=256, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 32188, Throughput in MB/s: 30338


* params:  n_rows=2000000, feat_size=256, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 73, Throughput in MB/s: 3


* params:  n_rows=2000000, feat_size=256, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 67, Throughput in MB/s: 3603


* params:  n_rows=2000000, feat_size=256, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 1205, Throughput in MB/s: 20247


* params:  n_rows=2000000, feat_size=256, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 8851, Throughput in MB/s: 27581


* params:  n_rows=2000000, feat_size=353, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 73, Throughput in MB/s: 18


* params:  n_rows=2000000, feat_size=353, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 113, Throughput in MB/s: 11894


* params:  n_rows=2000000, feat_size=353, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 5715, Throughput in MB/s: 23561


* params:  n_rows=2000000, feat_size=353, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 52995, Throughput in MB/s: 25409


* params:  n_rows=2000000, feat_size=353, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 75, Throughput in MB/s: 4


* params:  n_rows=2000000, feat_size=353, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 75, Throughput in MB/s: 4455


* params:  n_rows=2000000, feat_size=353, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 1822, Throughput in MB/s: 18468


* params:  n_rows=2000000, feat_size=353, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 15684, Throughput in MB/s: 21463


* params:  n_rows=20000000, feat_size=1, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 72, Throughput in MB/s: 0


* params:  n_rows=20000000, feat_size=1, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 68, Throughput in MB/s: 55


* params:  n_rows=20000000, feat_size=1, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 572, Throughput in MB/s: 666


* params:  n_rows=20000000, feat_size=1, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 2066, Throughput in MB/s: 1846


* params:  n_rows=20000000, feat_size=1, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 73, Throughput in MB/s: 0


* params:  n_rows=20000000, feat_size=1, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 68, Throughput in MB/s: 13


* params:  n_rows=20000000, feat_size=1, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 432, Throughput in MB/s: 220


* params:  n_rows=20000000, feat_size=1, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 1135, Throughput in MB/s: 839


* params:  n_rows=20000000, feat_size=4, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 72, Throughput in MB/s: 0


* params:  n_rows=20000000, feat_size=4, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 68, Throughput in MB/s: 223


* params:  n_rows=20000000, feat_size=4, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 630, Throughput in MB/s: 2419


* params:  n_rows=20000000, feat_size=4, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 3569, Throughput in MB/s: 4274


* params:  n_rows=20000000, feat_size=4, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 79, Throughput in MB/s: 0


* params:  n_rows=20000000, feat_size=4, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 70, Throughput in MB/s: 54


* params:  n_rows=20000000, feat_size=4, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 571, Throughput in MB/s: 666


* params:  n_rows=20000000, feat_size=4, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 2070, Throughput in MB/s: 1842


* params:  n_rows=20000000, feat_size=47, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 74, Throughput in MB/s: 2


* params:  n_rows=20000000, feat_size=47, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 80, Throughput in MB/s: 2220


* params:  n_rows=20000000, feat_size=47, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 1245, Throughput in MB/s: 14399


* params:  n_rows=20000000, feat_size=47, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 9343, Throughput in MB/s: 19188


* params:  n_rows=20000000, feat_size=47, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 73, Throughput in MB/s: 0


* params:  n_rows=20000000, feat_size=47, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 72, Throughput in MB/s: 620


* params:  n_rows=20000000, feat_size=47, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 692, Throughput in MB/s: 6469


* params:  n_rows=20000000, feat_size=47, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 4412, Throughput in MB/s: 10157


* params:  n_rows=20000000, feat_size=256, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 74, Throughput in MB/s: 13


* params:  n_rows=20000000, feat_size=256, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 96, Throughput in MB/s: 10160


* params:  n_rows=20000000, feat_size=256, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 4044, Throughput in MB/s: 24144


* params:  n_rows=20000000, feat_size=256, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 38530, Throughput in MB/s: 25345


* params:  n_rows=20000000, feat_size=256, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 75, Throughput in MB/s: 3


* params:  n_rows=20000000, feat_size=256, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 75, Throughput in MB/s: 3225


* params:  n_rows=20000000, feat_size=256, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 1812, Throughput in MB/s: 13472


* params:  n_rows=20000000, feat_size=256, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 9898, Throughput in MB/s: 24664


* params:  n_rows=20000000, feat_size=353, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 73, Throughput in MB/s: 18


* params:  n_rows=20000000, feat_size=353, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 118, Throughput in MB/s: 11381


* params:  n_rows=20000000, feat_size=353, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 5697, Throughput in MB/s: 23633


* params:  n_rows=20000000, feat_size=353, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 55568, Throughput in MB/s: 24232


* params:  n_rows=20000000, feat_size=353, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 77, Throughput in MB/s: 4


* params:  n_rows=20000000, feat_size=353, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 84, Throughput in MB/s: 3979


* params:  n_rows=20000000, feat_size=353, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 2089, Throughput in MB/s: 16110


* params:  n_rows=20000000, feat_size=353, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 15645, Throughput in MB/s: 21516


Total runtimes in us:  311991
```

After this PR with #6644:
```
* params:  dataset=ogbn-products n_rows=2449029 feat_size=100 dtype=torch.float32 num_indices=518134 feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 8758, Throughput in MB/s: 22566


* params:  n_rows=2000000, feat_size=1, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 64, Throughput in MB/s: 0


* params:  n_rows=2000000, feat_size=1, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 69, Throughput in MB/s: 55


* params:  n_rows=2000000, feat_size=1, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 245, Throughput in MB/s: 1555


* params:  n_rows=2000000, feat_size=1, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 471, Throughput in MB/s: 8090


* params:  n_rows=2000000, feat_size=1, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 69, Throughput in MB/s: 0


* params:  n_rows=2000000, feat_size=1, dtype=torch.int8, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 76, Throughput in MB/s: 12


* params:  n_rows=2000000, feat_size=1, dtype=torch.int8, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 146, Throughput in MB/s: 653


* params:  n_rows=2000000, feat_size=1, dtype=torch.int8, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 237, Throughput in MB/s: 4008


* params:  n_rows=2000000, feat_size=4, dtype=torch.float32, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 65, Throughput in MB/s: 0


* params:  n_rows=2000000, feat_size=4, dtype=torch.float32, num_indices=1000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 65, Throughput in MB/s: 234


* params:  n_rows=2000000, feat_size=4, dtype=torch.float32, num_indices=100000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 417, Throughput in MB/s: 3657


* params:  n_rows=2000000, feat_size=4, dtype=torch.float32, num_indices=1000000, feature_device=Device.Pinned, indices_device=Device.GPU

Runtime in us: 1124, Throughput in MB/s: 13572


* params:  n_rows=2000000, feat_size=4, dtype=torch.int8, num_indices=1, feature_device=Device.Pinned, indices_device=Device.GPU

…