@mfbalin 2023-11-28
@yaox12 how should Graphbolt depend on the thrust and cub in terms of the cmake setup?