@Rhett-Ying 2023-11-29
@mfbalin As for thrust, we're very like to merge https://github.com/dmlc/dgl/pull/6166 in a few days. If CCCL is ready, you need to include CCCL into GraphBolt?