@frozenbugs 2023-11-29
If this is a performance critical PR, we should make sure that it is fully tested.
Can you paste the result in the description of the PR?
I would expect, we test the overall index select speed given different reasonable sizes of features and reasonable numbers of ids. If we can cover multiple commonly used GPU, that would be a great plus.